feat: better windows support
Better support for Windows paths when used in cygwin or Git Bash. Note that you will have to npm
config set script-shell "C:\\path\\to\\bash.exe" to the bash executable that ships with either of
those. Also, yarn run is temporarily busted (they have a fix on master) when using that config so
you will need to use npm run in the meantime.